Updated screen output and source files to reflect version 2.0.

+++ garli/branches/partitioning/src/clamanager.h
@@ -1,5 +1,5 @@
-// GARLI version 0.96b8 source code
-// Copyright 2005-2008 Derrick J. Zwickl
+// GARLI version 2.0 source code
+// Copyright 2005-2011 Derrick J. Zwickl
 // email: zwickl@nescent.org
 //
 //  This program is free software: you can redistribute it and/or modify
@@ -191,7 +191,7 @@ class ClaManager{
 			if(holders[i].tempReserved ==true) tempres++;
 			if(holders[i].reserved==true) res++;
 			}		
-		assigned = numHolders - holderStack.size();
+		assigned = numHolders - (int) holderStack.size();
 		}
 	
 	inline int ClaManager::GetClaNumber(int index){
